Extract workspace_walk.rs from turn.rs (parent #680)
Move the workspace walker helpers out of turn.rs into a new
src/agent/loop_run/workspace_walk.rs sibling module:

- workspace_appears_empty (pub(super))
- meaningful_workspace_files (pub(super))
- collect_meaningful_workspace_files (private)

The walker honours `task_workspace_scope::is_workspace_ignored_dir` as
the SSOT ignore list. No facade re-export (DR3-001).

Callers updated:
- scaffold_pipeline.rs: switches `meaningful_workspace_files` and
  `workspace_appears_empty` imports from `super::turn` to
  `super::workspace_walk`.
- turn.rs: re-imports both via `super::workspace_walk::{...}` for the
  remaining in-file callers.

The legacy un-scoped `existing_workspace_candidate_for_role` (test-only)
and `existing_workspace_candidate_for_role_in_scope` are kept in turn.rs
for now since they depend on `scaffold_pipeline::scaffold_candidate_priority`
and `task_contract::ArtifactRole`; a follow-up PR can move them once a
cleaner module sibling is identified.

turn.rs: 26,781 -> 26,727 (-54 LOC)
Cumulative from 39,489: -12,762 LOC (-32.3%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
